requirements.txt
attrs==23.1.0
blinker==1.6.2
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
cryptography==41.0.1
exceptiongroup==1.1.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
kaitaistruct==0.10
outcome==1.2.0
pyasn1==0.5.0
pycparser==2.21
pyOpenSSL==23.2.0
pyparsing==3.1.0
PySocks==1.7.1
selenium==4.8.3
selenium-wire==5.1.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.1
trio-websocket==0.10.3
urllib3==1.26.16
wsproto==1.2.0
zstandard==0.21.0
pymongo==4.6.1
psycopg2-binary